Move Uncurrier out into its own phase
